Fixed long-standing >= bug in main_stream. Added extra casts in HASHTC for safety

+++ src/HASHTCounter.cpp
@@ -30,16 +30,16 @@ namespace HTC{
     void HASHTCounter::increment(htc_type key){
         //cout << (++counts [ key % my_size ]) << endl;
         #pragma omp atomic update
-        ++(counts[ key % my_size ]);
+        ++(counts[ key % (uint64_t) my_size ]);
     }
 
     int& HASHTCounter::get(htc_type key){
-        return (counts[ key % my_size ]);
+        return (counts[ key % (uint64_t) my_size ]);
     }
 
     void HASHTCounter::get(htc_type key, int& ret){
         #pragma omp atomic write 
-        ret = (counts[ key % my_size ]);
+        ret = (counts[ key % (uint64_t) my_size ]);
     }
 
     int HASHTCounter::size(void){
